Add C11 threading primitives for Fuchsia.
These are the preferred threading primitives on Fuchsia (v.s. POSIX
threads).

Currently this API is gated for #[cfg(target_os = "fuchsia")], because
support for C11 threads is on other platforms is limited and in some
cases requires particular versions of the OS/libc (e.g. FreeBSD 10.0+ or
musl libc 1.15+).